ny fordon class
